requirements.txt
requests==2.31.0
selenium==4.18.1
python-dotenv==1.0.0
